ORCH-1103: break MessageList<->ToolProposalCard require-cycle (move ConfirmOutcome to neutral toolProposalTypes module)
